Reworked matching algorithm
Fine tuning or rather a complete overhaul of how I determine the
quality of a guess to narrow the solution pool. Still not quite there
yet.

Was just matching exact matches but with each guess it'll would only
match the latest guess. A good guess should have the same number of
exact matches or at least equal inexact. (I need to think this through
a bit more)

Overall flow of the program was reworked as well.